Uselles files deleted

+++ Fish On The Sea/src/Screens/Gameplay/gameplay.cpp
@@ -446,7 +446,7 @@ namespace fish {
 				//EndMode2D();
 				break;
 			case GameStage::Pause:
-				DrawTexture(bigBox, boxPosX, boxPosY, WHITE);
+				DrawTexture(bigBox, static_cast<int>(boxPosX), static_cast<int>(boxPosY), WHITE);
 				DrawTextEx(font, "Press P ro go back to the game", { 23,70 }, 24, 3, BLACK);
 				DrawTextEx(font, "Press Menu to go back and lose progress", { 23,95 }, 24, 3, BLACK);
 				DrawTextEx(font, "Press the sound icon to mute or un mute", { 23,140 }, 24, 3, BLACK);
